% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threejstable.R
\name{threejstable-shiny}
\alias{threejstable-shiny}
\alias{threejstableOutput}
\alias{renderThreejstable}
\title{Shiny bindings for threejstable}
\usage{
threejstableOutput(outputId, width = "100\%", height = "400px")

renderThreejstable(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{expr}{An expression that generates a threejstable}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Output and render functions for using threejstable within Shiny
applications and interactive Rmd documents.
}
